initialize allocator on the controlplane instance
The allocators for Services ClusterIPs and NodePorts requires to
share information between the controlplane and the Rest storage.

Instead of initializing them on the rest storage, and sharing them
with the controlplane instance, do the other way around, initialize
them on the controlplane instance and share them with the rest storage.
